Fix plot issues

+++ Telemetry/DataLayer/Models/Speed.cs
@@ -1,7 +1,7 @@
 ﻿namespace DataLayer.Models
 {
-    public class Speed : Sensor<int>
+    public class Speed : Sensor<float>
     {
-        public override int Value { get => base.Value; set => base.Value = value; }
+        public override float Value { get => base.Value; set => base.Value = value; }
     }
 }
